src/app/requirements.txt
fastapi  
uvicorn  
azure-cognitiveservices-speech  
openai  
numpy  
opencv-python-headless
jinja2 
python-dotenv
requests